arabic-writing-practice-app/requirements.txt
streamlit>=1.22.0
transformers>=4.28.0
torch>=2.0.0
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.7.0
python-bidi>=0.4.2
streamlit-drawable-canvas>=0.9.2
scikit-learn>=1.2.2
arabic-reshaper>=3.0.0
pytesseract>=0.3.10
difflib>=0.3.6
arabic-reshaper>=3.0.0
streamlit-drawable-canvas>=0.9.2

